Add PFR capabilities to mtd-util
When PFR is in pass-thru mode (non-enforcing), the BMC needs to be able
to consume a signed bmc update in PFR format and write it to the active
region. This change adds the extraction from a signed, compressed image
to the active region.

Tested: performed a manual firmware update via mtd-util p w <img.bin>
        Also tested using the emulation mode and bindiff with the mtd
        image.

Change-Id: I84775a40ce10a193c143b1c0f1414ac659afbb07
Signed-off-by: Vernon Mauery <vernon.mauery@intel.com>