Add option to force factory reset when doing pfr write
PFR write from a non-pfr system to a PFR system needs to force the erase
regions to be cleared or the unsigned regions of the flash may end up
with bits that should not be there. This makes it so that the fwupd.sh
script can detect that situation and force a recovery.

Tested: 1) Build non-PFR image; flash to BMC
        2) Build PFR image
        3) Update non-PFR BMC using PFR image
        4) watch that the NV regions are cleared and no JFFS2 errors
           are printed by the kernel at the next boot
        5) Tested PFR provisioning after migrating from non-PFR
           to PFR.

Change-Id: Ifb0a9b0307d81e197c05b08688b5293284be449f
Signed-off-by: Vernon Mauery <vernon.mauery@intel.com>
Signed-off-by: Vikram Bodireddy <vikram.bodireddy@intel.com>