Fix erase alignment error during firmware update
Issue found on some system that OOB firmware update was bricking
the system after the update. The root cause is as oob update's
flash erase regions aren't aligned to the flash write regions.
Flash erase logic is erasing only 64KB aligned pages so that
it's missing the other 4K incremental flash write pages.
Fix will cross check if the write region was erased part of 64KB
sector erase previously or not and if not erase it before the write.

Tested: tested OOB update on failed system complete successfully.

Change-Id: I3af18c95b5bf1d0ce02dc87bb237c84107654205
Signed-off-by: Vikram Bodireddy <vikram.bodireddy@intel.com>