Implement FVM authentication
For capsules that contain Firmware Volume Manifests (FVMs),
mtd-util will now parse the authentication headers inside of the signed
external capsule signature.

Sha256 and ecdsa256 are deprecated for products going forward. This
removes them from the authentication path, rejecting images that attempt
to use them and enforcing an empty hash in the b0 sha256 hash location.

Check the FITc layout ID in the FVM header against the layout ID
that was stored previously. If no ID was stored, the default is 0, which
is also the default in images that don't have one set.

Tested: built native version and authenticated new partial capsules
        built for target and performed  update

        will reject images with the sha256-present bit set in a region
        and will warn on images that contain non-ff bytes in the global
        sha256 hash region.

        run mtd-util pfr authenticate <image> to see that it rejects
        images with a different layout ID than the filesystem.

Change-Id: I82ee92932291e82859a2a1afbdc199ab30b14fa4
Signed-off-by: Vernon Mauery <vernon.mauery@intel.com>