Enforce no sha256 hashing present
sha256 is no longer allowed. Prior to this, it was ignored if present
to allow for dual-hashed sha256/sha384 images to pass. Now fully remove
the support for sha256, requiring the sha256_present bit to be 0 and
requiring that the sha256 area in b0 is filled with 0xff.

Change-Id: Ie686d27325a8ff5f2a636a6c94ef363aada7fb77
Tested: checked that sha256 images are rejected
Signed-off-by: Vernon Mauery <vernon.mauery@intel.com>